profile page redux done
