+++ src/app/store.js
@@ -1,10 +1,12 @@
 import { combineReducers, createStore } from "redux";
 import { default as themeReducer } from "../features/theme/theme.reducer";
 import { default as freelancesReducer } from '../features/freelances/freelances.reducer'
+import { default as freelanceReducer } from '../features/freelance/freelance.reducer'
 
 let reducer = combineReducers({
     theme: themeReducer,
-    freelances: freelancesReducer
+    freelances: freelancesReducer,
+    freelance: freelanceReducer
 })
 
 //debug tool
